read images from firestore
